Setup for extension to 2D
This is the conservative spectral code for modelling the inhomegenous Landau-Poisson equation, still running in one space dimension, but with the arrays extended and indicies shifted, ready for extension to two space dimensions.  Note that so far there is only focus on running the (collisionless) Vlasov problem.  The results match with the usual code for one space dimension.