Added entropy feature
The EntropyCalculation files have been added to calculate the entropy at each time-step.  Note that the function to compute relative entropy is commented out because it would require an equilibrium solution which has not been computed yet for this code.  This is still just code for calculating entropy in the 1D simulation, but the indices have been shifted appropriately.  The values computed agree with the previous 1D simulations as well.