Changed U & U1 from datatype double* to vector<double>
The datatypes of U & U1 have been changed from double* (a C construct) to vector<double> (a C++ construct) for the sake of efficiency, and also because the code from Jose's Poisson solver was adapted to handle the vector<double> datatype.  Note that all functions using U & U1 had their corresponding arguments altered to reflect this change.  All values seem to be the same, except for at machine accuracy in the momentum calculations, which should not be an issue.