Added Jose's Poisson code into Chenglong's code (as a check - not used in advection yet)
The source code for Jose's 2D LDG solver of Poisson's equation has been added (namely the PoissonVariables.cpp, PoissonFunctions.cpp & PoissonPrint.cpp files and related header files) and plots produced to prove that it does indeed give the correct solution for the Landau damping initial condition.  In order for this to be done, the initial conditions had to be updated in the SetInit_1.cpp file to give a full set of coefficients for 2D.  In the process, the MarginalCreation files were updated to now produce two different marginal plots, one in the variables (x1,v1) and another in the variables (x1,x2).  The main file was editted accordingly to produce the data for these plots.  Also, in order to be compatible with the code for the Poisson solver, the C++ vector class is now being used (for DG coefficients and field data, etc.).