Updated I2 in advection.cpp to the 2D version as well as adding Int_fE1 & Int_fE2 - still giving similar 1D results
The function I2 has been updated with the results of the 2D DG calculations, but still need to update the system to solve for H.  Note that the Int_fE function had to be replaced with Int_fE1 & Int_fE2 functions which now both use the 2D Poisson solver integrals.  This is causing an error in the conservation of the v1 component of momentum (definitely the values from IXE_X, not the new I2 formulation).