Fixed specular reflection BC formulation
The original implementation of the specular reflection BCs was far too naive.  The correct formulation has now been implemented and the code is conserving as it should.  There is still an error when the field is introduced in the first component of momentum as well as energy, but with no field everything is conserved perfectly in the collisionless case.